ts-ify handlePreviewToggles
